rel 1.1.5 bug 3.79

+++ app/src/main/java/kr/sweetapps/alcoholictimer/ui/tab_02/screens/AddRecordScreen.kt
@@ -415,7 +415,7 @@ fun AddRecordScreenComposable(
                                         startTime = startMillis,
                                         endTime = endMillis,
                                         targetDays = commitTarget,
-                                        actualDays = actualDays,
+                                        actualDays = actualDays.toDouble(),  // [FIX] Int → Double 변환
                                         isCompleted = completed,
                                         status = status,
                                         createdAt = System.currentTimeMillis()
